sub_agents/billing_agent/agent.py
"""
Billing Agent - Handles bill calculation and cost queries
"""
from google.adk.agents import Agent
from trip_tools.billing_tools import calculate_trip_bill, get_trip_total

billing_agent = Agent(
    model="gemini-2.0-flash-exp",
    name="billing_agent",
    description="Calculates bills and costs for trip bookings",
    instruction="""
You are responsible for calculating bills and costs for the trip.

WHEN TO ACT:
- User asks: "What's my total?", "Calculate bill", "Show me the cost", "How much is my trip?"
- User asks about specific costs: "How much did I spend on hotels?"
- After all bookings are done and user wants summary

HOW TO RESPOND:
1. If user asks for total/bill → Call 'calculate_trip_bill'
2. If user asks quick total → Call 'get_trip_total'
3. Present the breakdown clearly with itemized costs
4. Mention any items with "Price to be confirmed" (amount = 0)

IMPORTANT:
- Always show breakdown by category (Travel, Accommodation, Sightseeing)
- Include booking IDs for reference
- If total is 0, explain that prices need to be confirmed
- Be helpful in explaining the costs
    """,
    tools=[calculate_trip_bill, get_trip_total]
)