requirements.txt
google-adk==1.3.0
google-generativeai==0.5.4
python-dotenv==1.0.1
tinydb==4.8.0
loguru==0.7.2
pytest==8.2.1
aiosqlite==0.19.0
twilio>=8.10.0